refactor(reset-tokens): refactored reset password tokens services to meet new token business rules
fixes failing tests at PR #63
