+++ .env.example
@@ -6,8 +6,7 @@ JWT_SECRET=some_secret_hash
 
 ACCESS_TOKEN_EXPIRES_IN=24h
 REFRESH_TOKEN_EXPIRES_IN=7d
-RESET_TOKEN_EXPIRES_TIME=30
-RESET_TOKEN_EXPIRES_TIME_FORMAT=minutes
+RESET_TOKEN_EXPIRES_IN=30m
 
 DB_NAME=node-bloodboiler-development
 DB_HOST=localhost
