+++ src/services/auth/resetPassword.service.js
@@ -1,20 +1,32 @@
 const { StatusCodes } = require('http-status-codes');
-const { usersRepository } = require('../../repositories');
+
 const { ApplicationError, jwt } = require('../../utils');
 const { messages } = require('../../helpers');
-const userService = require('../users/update.service');
+
+const userService = require('../users');
+const { tokensRepository } = require('../../repositories');
+const { TokenTypes } = require('../../models');
 
 module.exports.resetPassword = async (token, newPassword) => {
-  const user = await usersRepository.get({ passwordResetToken: token });
-  if (!user) {
-    throw new ApplicationError(messages.notFound('user'), StatusCodes.NOT_FOUND);
+  const resetToken = await tokensRepository.get({ token, tokenType: TokenTypes.reset });
+
+  if (!resetToken) {
+    throw new ApplicationError(messages.notFound('token'), StatusCodes.NOT_FOUND);
   }
 
-  jwt.verify(token, (err) => {
-    if (err) {
-      throw new ApplicationError(messages.expiredToken, StatusCodes.UNAUTHORIZED);
-    }
-  });
+  const decoded = await jwt.verifyToken(token);
 
-  await userService.update(user._id, { password: newPassword, passwordResetToken: null });
+  if (decoded.expired) {
+    resetToken.hasExpired = true;
+    await tokensRepository.update(resetToken);
+  }
+
+  if (resetToken.hasExpired) {
+    throw new ApplicationError(messages.expiredToken, StatusCodes.FORBIDDEN);
+  }
+
+  await userService.update(decoded.sub.id, { password: newPassword });
+
+  resetToken.hasExpired = true;
+  await tokensRepository.update(resetToken);
 };
